+++ code/models.py
@@ -2,9 +2,10 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-class Net(nn.Module):
+# TODO rajouter le get_features dans l'énoncé
+class MNISTNet(nn.Module):
     def __init__(self):
-        super(Net, self).__init__()
+        super(MNISTNet, self).__init__()
         self.conv1 = nn.Conv2d(1, 6, 5)
         self.pool = nn.MaxPool2d(2, 2)
         self.conv2 = nn.Conv2d(6, 16, 5)
@@ -13,9 +14,11 @@ class Net(nn.Module):
         self.fc3 = nn.Linear(84, 10)
 
     def forward(self, x):
-        x = self.pool(F.relu(self.conv1(x)))
-        x = self.pool(F.relu(self.conv2(x)))
-        x = x.view(-1, 16 * 4 * 4)
+        x = F.relu(self.conv1(x))
+        x = torch.max_pool2d(x, (2,2))
+        x = F.relu(self.conv2(x))
+        x = torch.max_pool2d(x, (2,2))
+        x = torch.flatten(x, 1) #or x.view(-1, 16 * 4 * 4)
         x = F.relu(self.fc1(x))
         x = F.relu(self.fc2(x))
         x = self.fc3(x)
@@ -27,70 +30,9 @@ class Net(nn.Module):
         x = x.view(-1, 16 * 4 * 4)
         return x
 
-def double_conv(in_channels, out_channels):
-    return nn.Sequential(
-        nn.Conv2d(in_channels, out_channels, 3, padding=1),
-        nn.ReLU(inplace=True),
-        nn.Conv2d(out_channels, out_channels, 3, padding=1),
-        nn.ReLU(inplace=True)
-    )   
 
-class UNet(nn.Module):
-
-    def __init__(self):
-        super().__init__()
-                
-        self.dconv_down1 = double_conv(1, 64)
-        self.dconv_down2 = double_conv(64, 128)
-        self.dconv_down3 = double_conv(128, 256)
-        self.dconv_down4 = double_conv(256, 512)        
-
-        self.maxpool = nn.MaxPool2d(2)
-        self.upsample = nn.Upsample(scale_factor=2, mode='bilinear', align_corners=True)        
-        
-        self.dconv_up3 = double_conv(256 + 512, 256)
-        self.dconv_up2 = double_conv(128 + 256, 128)
-        self.dconv_up1 = double_conv(128 + 64, 64)
-        
-        self.conv_last = nn.Conv2d(64, 3, 1)
-        
-        
-    def forward(self, x):
-        conv1 = self.dconv_down1(x)
-        x = self.maxpool(conv1)
-
-        conv2 = self.dconv_down2(x)
-        x = self.maxpool(conv2)
-        
-        conv3 = self.dconv_down3(x)
-        x = self.maxpool(conv3)   
-        
-        x = self.dconv_down4(x)
-        
-        x = self.upsample(x)        
-        x = torch.cat([x, conv3], dim=1)
-        
-        x = self.dconv_up3(x)
-        x = self.upsample(x)        
-        x = torch.cat([x, conv2], dim=1)       
-
-        x = self.dconv_up2(x)
-        x = self.upsample(x)        
-        x = torch.cat([x, conv1], dim=1)   
-        
-        x = self.dconv_up1(x)
-        
-        out = self.conv_last(x)
-        
-        return out
-
-    def get_features(self, x):
-            conv1 = self.dconv_down1(x)
-            x = self.maxpool(conv1)
-
-            conv2 = self.dconv_down2(x)
-            x = self.maxpool(conv2)
-            
-            conv3 = self.dconv_down3(x)
-            x = self.maxpool(conv3)
-            return x
+if __name__=='__main__':
+    x = torch.rand(16,1,28,28)
+    net = MNISTNet()
+    y = net(x)
+    assert y.shape == (16,10)
